feat: Implement Profile Repository
Adds the basic structure for a profile repository (`ProfileRepo`) and its implementation (`ProfileRepoImp`).

The implementation uses the `IApiService` to fetch user data by ID from the endpoint `user/user001`. It handles the API response, converts the data to a list of `UserModel`, and returns it within a `Either` type to signify success or failure.

Also, slightly refactors `PlaceModel` and `Location` for better readability and adds a print statement in `CoreApiService` for debugging.